Add files via upload
Project Overview: Superstore Analysis
In this data analysis project, I showcased my proficiency in data manipulation, cleaning, and visualization techniques by conducting a thorough examination of the Superstore dataset. The project not only explores the inherent structure of the data but also uncovers valuable insights into various aspects of the business, making it a robust addition to my portfolio.
•	DataSetup:
The project starts with the importation of crucial libraries, laying the foundation for effective data analysis and visualization.
Display options are configured for Pandas to enhance data exploration.
•	Data Loading and Exploration:
The Superstore dataset is loaded into a Pandas DataFrame, and its shape is examined to understand the overall structure.
A meticulous approach is taken to identify and remove duplicate rows, ensuring the integrity of the dataset.
•	Data Cleaning:
A strategic decision is made to optimize memory usage by converting the data type of the 'Segment' column to a category.
•	Data Visualization:

Leveraging Seaborn, I crafted a visually compelling count plot, shedding light on the distribution of sub-categories within the Superstore dataset.
•	Product Analysis:

A deeper dive into the dataset involves grouping data by category and product name, resulting in a pie chart that effectively communicates the sales distribution of top products.
•	Region and Category Distribution:

The project delves into regional preferences by generating a comprehensive count plot that visually captures the distribution of product categories across different regions.
•	Monthly Sales Analysis:
One of the project highlights is the exploration of monthly sales patterns. This involved calculating and visualizing the average discount, quantity, and profit across different months, providing valuable insights into seasonal trends.